git-svn-id: svn://cherokee-project.com/cherokee/trunk@528 5dc97367-97f1-0310-9951-d761b3857238

+++ cherokee/handler_fastcgi.h
@@ -63,7 +63,7 @@ typedef struct {
 	cuint_t                         id;
 	cuchar_t                        generation;	
 	cherokee_buffer_t               write_buffer;	
-	size_t                          post_len;
+	off_t                           post_len;
 
 	cherokee_fcgi_manager_t        *manager;
 	cherokee_fcgi_dispatcher_t     *dispatcher;
